lore/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.0
google-genai==1.0.0
google-cloud-aiplatform==1.60.0
pypdf==4.3.1
pydantic==2.8.0
python-multipart==0.0.9
google-api-python-client==2.140.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
